requirements.txt
pymongo
feedparser
python-telegram-bot
wordcloud
python-bidi